Refactor method names for clarity and consistency in indentation handling

+++ test/ShapeCrawler.Tests.Unit/FontTests.cs
@@ -74,7 +74,7 @@ public class FontTests : SCTest
     [SlideShape("014.pptx", 5, 4, 12)]
     [SlideShape("014.pptx", 6, 52, 27)]
     [SlideShape("autoshape-case016.pptx", 1, "Text Placeholder 1", 28)]
-    public void Size_Getter_returns_font_size(IShape shape, decimal expectedSize)
+    public void Size_Getter_returns_font_size(IShape shape, double expectedSize)
     {
         // Arrange
         var autoShape =  shape;
@@ -84,7 +84,7 @@ public class FontTests : SCTest
         var fontSize = font!.Size;
         
         // Assert
-        fontSize.Should().Be(expectedSize);
+        fontSize.Should().Be((decimal)expectedSize);
     }
     
     [Test]
